rename string_table_builder_ module
